isolating auth config to separate file
